requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.2
cycler==0.12.1
distro==1.9.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.13
fastapi-cli==0.0.7
einops==0.8.0
filelock==3.18.0
fire==0.7.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
fugashi==1.5.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.0
idna==3.10
jaconv==0.4.0
Jinja2==3.1.6
jiter==0.10.0
kiwisolver==1.4.8
kornia==0.7.4
loguru==0.7.3
manga-ocr==0.1.14
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.5.0
networkx==3.5
numpy==2.3.0
openai==1.90.0
opencv-python==4.11.0.86
packaging==25.0
pandas==2.3.0
pillow==11.2.1
propcache==0.3.2
psutil==7.0.0
py-cpuinfo==9.0.0
pydensecrf>=1.0.0
pyclipper>=1.3.0
shapely>=2.0.0
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rich-toolkit==0.14.7
safetensors==0.5.3
scipy==1.15.3
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
termcolor==3.1.0
tokenizers==0.21.1
torch==2.7.1+cu128
torchvision==0.22.1+cu128
tqdm==4.67.1
transformers==4.52.4
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
ultralytics==8.4.14
ultralytics-thop==2.0.14
unidic-lite==1.0.8
urllib3==2.5.0
uvicorn==0.34.3
watchfiles==1.1.0
websockets==15.0.1
wheel==0.45.1
win32_setctime==1.2.0
yarl==1.20.1

# MIT 渲染模块依赖
PyQt6>=6.6.0
PyHyphen>=4.0.4
langcodes>=3.4.0
pythainlp>=5.0.0
